% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSUN.R
\name{csun_thesis}
\alias{csun_thesis}
\title{CSUN Thesis Template}
\usage{
csun_thesis(..., keep_tex = TRUE, pandoc_args = c("--chapters"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Write your thesis for CSUN in R Markdown
}
\references{
href{https://www.sharelatex.com/templates/thesis/csun-thesis}
}
\author{
Carl Ganz
}
